+++ app/trace/t_test.go
@@ -1,0 +1,21 @@
+package trace
+
+import (
+	"bytes"
+	"testing"
+)
+
+func TestNew(t *testing.T) {
+	var buf bytes.Buffer
+	tracer := New(&buf)
+	if tracer == nil {
+		t.Error("Newからの戻り値がnilです")
+	} else {
+		tracer.Trace("こんにちは、traceパッケージ")
+		if buf.String() != "こんにちは、traceパッケージ\n" {
+			t.Errorf("'%s'という誤った文字列が出力されました", buf.String())
+		}
+	}
+}
+
+
