+++ app/trace/tracer.go
@@ -1,0 +1,30 @@
+package trace
+
+import (
+	"fmt"
+	"io"
+)
+
+type Tracer interface {
+	Trace(...interface{})
+}
+
+type tracer struct {
+	out io.Writer
+}
+
+type nilTracer struct{}
+
+func (t *nilTracer) Trace(a ...interface{}) {}
+func Off() Tracer {
+	return &nilTracer{}
+}
+
+func (t *tracer) Trace(a ...interface{}) {
+	t.out.Write([]byte(fmt.Sprint(a...)))
+	t.out.Write([]byte("\n"))
+}
+
+func New(w io.Writer) Tracer {
+	return &tracer{out: w}
+}
