+++ app/views/css/chat.css
@@ -1,0 +1,8 @@
+#chatbox {
+  position: fixed;
+  width: 85%;
+  bottom: 1%;
+}
+#chatbox input[type="text"] {
+  width: 80%;
+}
