+++ app/views/templates/js/index.html
@@ -1,0 +1,3 @@
+{{define "script" }}
+<!-- 現状なし -->
+{{end}}
