--- chat/room.go
@@ -1,5 +1,0 @@
-package chat
-
-type room struct {
-	forward chan []byte //メッセージの送信に使う
-}
